refactoring tests + transpiler cont.
