+++ qbraid/programs/pyquil.py
@@ -62,13 +62,13 @@ class PyQuilProgram(QuantumProgram):
     def _contiguous_expansion(self) -> None:
         """Checks whether the circuit uses contiguous qubits/indices,
         and if not, adds identity gates to vacant registers as needed."""
-        return NotImplementedError
+        raise NotImplementedError
 
     def _contiguous_compression(self) -> None:
         """Checks whether the circuit uses contiguous qubits/indices,
         and if not, reduces dimension accordingly."""
-        return NotImplementedError
+        raise NotImplementedError
 
     def reverse_qubit_order(self) -> None:
         """Reverse the order of the qubits in the circuit."""
-        return NotImplementedError
+        raise NotImplementedError
