+++ qbraid/transpiler/cirq_qiskit/conversions.py
@@ -17,7 +17,6 @@ import cirq
 import qiskit
 
 from qbraid import circuit_wrapper
-from qbraid.interface.qbraid_cirq.tools import _convert_to_line_qubits
 from qbraid.transpiler.cirq_qasm import from_qasm, to_qasm
 from qbraid.transpiler.custom_gates import _map_zpow_and_unroll
 from qbraid.transpiler.exceptions import CircuitConversionError
@@ -56,6 +55,8 @@ def from_qiskit(circuit: qiskit.QuantumCircuit) -> cirq.Circuit:
     try:
         qasm_str = circuit.qasm()
         cirq_circuit = from_qasm(qasm_str)
-        return _convert_to_line_qubits(cirq_circuit)
+        qprogram = circuit_wrapper(cirq_circuit)
+        qprogram._convert_to_line_qubits()
+        return qprogram.program
     except Exception as err:
         raise CircuitConversionError from err
