+++ tests/interface/test_braket_qasm_conversions.py
@@ -22,6 +22,16 @@ def test_braket_to_qasm3_bell_circuit():
     """Test converting braket bell circuit to OpenQASM 3.0 string"""
     qasm_expected = """
 OPENQASM 3.0;
+bit[2] b;
+qubit[2] q;
+h q[0];
+cnot q[0], q[1];
+b[0] = measure q[0];
+b[1] = measure q[1];
+"""
+
+    qasm_expected_2 = """
+OPENQASM 3.0;
 bit[2] __bits__;
 qubit[2] __qubits__;
 h __qubits__[0];
@@ -30,7 +40,8 @@ __bits__[0] = measure __qubits__[0];
 __bits__[1] = measure __qubits__[1];
 """
     bell = Circuit().h(0).cnot(0, 1)
-    assert qasm_expected.strip("\n") == braket_to_qasm3(bell)
+    qbraid_qasm = braket_to_qasm3(bell)
+    assert qasm_expected.strip("\n") == qbraid_qasm or qasm_expected_2.strip("\n") == qbraid_qasm
 
 
 def test_braket_from_qasm3():
