+++ tests/interface/test_programs.py
@@ -16,8 +16,8 @@ import pytest
 
 from qbraid import circuit_wrapper
 from qbraid.exceptions import ProgramTypeError, VisualizationError
-from qbraid.interface.calculate_unitary import circuits_allclose
 from qbraid.interface.circuit_drawer import circuit_drawer
+from qbraid.interface.circuit_equality import circuits_allclose
 from qbraid.interface.random_circuit import random_circuit
 
 from .._data.programs import bell_data, shared15_data
