+++ tests/transpiler/cirq_pytket/test_conversions_pytket.py
@@ -19,7 +19,7 @@ from pytket.circuit import Circuit as TKCircuit
 from pytket.qasm import circuit_to_qasm_str
 
 from qbraid.interface import circuits_allclose
-from qbraid.interface.qbraid_cirq.tools import _equal
+from qbraid.interface.circuit_equality import _equal
 from qbraid.transpiler.cirq_pytket.conversions import from_pytket, to_pytket
 from qbraid.transpiler.cirq_qasm import from_qasm
